Added permission checks
I realized i didnt have any checks so i was like "I should add them

i havent tested all of them, but they should work in theory

+++ commands/purge.js
@@ -1,12 +1,18 @@
 exports.run = function(client, message, args) {
-  let messagecount = parseInt(args.join(' '));
-  message.channel.fetchMessages({
-    limit: messagecount
-  }).then(messages => {
-    message.channel.bulkDelete(messages)
-    message.channel.send(`Deleted ${messagecount} messages.`)
-    .then(delmsg => {
-      delmsg.delete(10000)
-    })
-  });
+  let permissionCheck = message.author.id
+  if (message.guild.member(permissionCheck).hasPermission("MANAGE_MESSAGES")) {
+    let messagecount = parseInt(args.join(' '));
+    message.channel.fetchMessages({
+      limit: messagecount
+    }).then(messages => {
+      message.channel.bulkDelete(messages)
+      message.channel.send(`Deleted ${messagecount} messages.`)
+      .then(delmsg => {
+        delmsg.delete(10000)
+      })
+    });
+  }
+  else {
+    message.channel.send("`purge.exe has stopped working`: Missing permissions")
+  }
 };
